функция отрисовки игровых экранов вынесена в utils
