+++ js/two-cards-game-screen.js
@@ -1,11 +1,7 @@
-import {getElementFromTemplate, getLevelProgressBar} from './utils.js';
-import {showScreen} from './main.js';
-import {getOneCardGameScreen} from './one-card-game-screen.js';
-import {getThreeCardsGameScreen} from './three-cards-game-screen.js';
+import {getElementFromTemplate, getLevelProgressBar, enterNextLevel} from './utils.js';
 import getHeader from './header.js';
-import getFooter from './footer.js';
-import getStatsScreenElement from './stats-screen.js';
-import {levels, answers, userState} from './data.js';
+import getFooterMarkup from './footer.js';
+import {answers, userState} from './data.js';
 
 const addLevelResult = (data, firstCardInputsValue, secondCardInputsValue) => {
   let answerOnCard = {};
@@ -21,22 +17,8 @@ const addLevelResult = (data, firstCardInputsValue, secondCardInputsValue) => {
       solved: false
     };
   }
+  answers.push(answerOnCard);
   return answerOnCard;
-};
-
-const checkSelectedAnswer = (data, firstCardInputsValue, secondCardInputsValue) => {
-  const numberOfScreen = Array.prototype.indexOf.call(levels, data);
-  const nextLevel = levels[numberOfScreen + 1];
-  answers.push(addLevelResult(data, firstCardInputsValue, secondCardInputsValue));
-  if (nextLevel && userState.lives > 0) {
-    if (nextLevel.levelType === `one-card`) {
-      showScreen(getOneCardGameScreen(nextLevel, userState));
-    } else if (nextLevel.levelType === `three-cards`) {
-      showScreen(getThreeCardsGameScreen(nextLevel, userState));
-    }
-  } else {
-    showScreen(getStatsScreenElement(answers, userState));
-  }
 };
 
 const template = (level) => `
@@ -64,7 +46,7 @@ const template = (level) => `
       </ul>
     </div>
   </div>
-  ${getFooter()}
+  ${getFooterMarkup()}
   </main>
   `;
 
@@ -76,7 +58,8 @@ const getTwoCardsGameScreen = (data, state) => {
   const secondCardRadioInputs = form.elements.question2;
   form.addEventListener(`change`, function () {
     if (firstCardRadioInputs.value && secondCardRadioInputs.value) {
-      checkSelectedAnswer(data, firstCardRadioInputs.value, secondCardRadioInputs.value);
+      addLevelResult(data, firstCardRadioInputs.value, secondCardRadioInputs.value);
+      enterNextLevel(data, firstCardRadioInputs.value, secondCardRadioInputs.value);
     }
   });
   return el;
